[JSC] Make StaticStringImpl & StaticSymbolImpl actually static https://bugs.webkit.org/show_bug.cgi?id=194212
Reviewed by Mark Lam.

Avoid mutation onto refcounts if `isStatic()` returns true so that the content of StaticStringImpl never gets modified.

* wtf/text/StringImpl.h:
(WTF::StringImpl::ref):
(WTF::StringImpl::deref):

Canonical link: https://commits.webkit.org/209879@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@242747 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WTF/wtf/text/StringImpl.h
@@ -1064,6 +1064,8 @@ inline void StringImpl::ref()
 {
     STRING_STATS_REF_STRING(*this);
 
+    if (UNLIKELY(isStatic()))
+        return;
     m_refCount += s_refCountIncrement;
 }
 
@@ -1071,6 +1073,8 @@ inline void StringImpl::deref()
 {
     STRING_STATS_DEREF_STRING(*this);
 
+    if (UNLIKELY(isStatic()))
+        return;
     unsigned tempRefCount = m_refCount - s_refCountIncrement;
     if (!tempRefCount) {
         StringImpl::destroy(this);
